feat: reorganize operations module to support both op return and commit reveal schemes
