+++ sbtc-core/src/operations/commit_reveal/mod.rs
@@ -1,2 +1,6 @@
-pub mod construction;
+/*!
+Primitives for sBTC commit reveal transactions
+*/
+pub mod deposit;
 pub mod utils;
+pub mod withdrawal_request;
