+++ sbtc-core/src/operations/commit_reveal/withdrawal_request.rs
@@ -1,5 +1,5 @@
 /*!
-Construction of commit reveal transactions
+Primitives for sBTC commit reveal withdrawal request transactions
 */
 use std::io;
 
@@ -7,53 +7,12 @@ use bdk::bitcoin::{
     secp256k1::ecdsa::RecoverableSignature, Address as BitcoinAddress, Amount, Transaction, TxOut,
     XOnlyPublicKey,
 };
-use stacks_core::{codec::Codec, utils::PrincipalData};
+use stacks_core::codec::Codec;
 
 use crate::operations::{
     commit_reveal::utils::{commit, reveal, CommitRevealResult, RevealInputs},
     Opcode,
 };
-
-/// Data to construct a commit reveal deposit transaction
-pub struct DepositData {
-    /// Address or contract to deposit to
-    pub principal: PrincipalData,
-    /// How much to send for the reveal fee
-    pub reveal_fee: Amount,
-}
-
-impl Codec for DepositData {
-    fn codec_serialize<W: io::Write>(&self, dest: &mut W) -> io::Result<()> {
-        Codec::codec_serialize(&Opcode::Deposit, dest)?;
-        self.principal.codec_serialize(dest)?;
-        self.reveal_fee.codec_serialize(dest)?;
-
-        todo!()
-    }
-
-    fn codec_deserialize<R: io::Read>(data: &mut R) -> io::Result<Self>
-    where
-        Self: Sized,
-    {
-        let opcode = Opcode::codec_deserialize(data)
-            .map_err(|err| io::Error::new(io::ErrorKind::InvalidData, err))?;
-
-        if !matches!(opcode, Opcode::Deposit) {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidData,
-                "Invalid opcode, expected deposit",
-            ));
-        }
-
-        let principal = PrincipalData::codec_deserialize(data)?;
-        let reveal_fee = Amount::codec_deserialize(data)?;
-
-        Ok(Self {
-            principal,
-            reveal_fee,
-        })
-    }
-}
 
 /// Data to construct a commit reveal withdrawal transaction
 pub struct WithdrawalData {
@@ -67,6 +26,7 @@ pub struct WithdrawalData {
 
 impl Codec for WithdrawalData {
     fn codec_serialize<W: io::Write>(&self, dest: &mut W) -> io::Result<()> {
+        Codec::codec_serialize(&Opcode::WithdrawalRequest, dest)?;
         self.amount.codec_serialize(dest)?;
         self.signature.codec_serialize(dest)?;
         self.reveal_fee.codec_serialize(dest)
@@ -76,6 +36,16 @@ impl Codec for WithdrawalData {
     where
         Self: Sized,
     {
+        let opcode = Opcode::codec_deserialize(data)
+            .map_err(|err| io::Error::new(io::ErrorKind::InvalidData, err))?;
+
+        if !matches!(opcode, Opcode::WithdrawalRequest) {
+            return Err(io::Error::new(
+                io::ErrorKind::InvalidData,
+                "Invalid opcode, expected withdrawal request",
+            ));
+        }
+
         let amount = Amount::codec_deserialize(data)?;
         let signature = RecoverableSignature::codec_deserialize(data)?;
         let reveal_fee = Amount::codec_deserialize(data)?;
@@ -86,15 +56,6 @@ impl Codec for WithdrawalData {
             reveal_fee,
         })
     }
-}
-
-/// Constructs a peg in payment address
-pub fn deposit_commit(
-    deposit_data: DepositData,
-    revealer_key: &XOnlyPublicKey,
-    reclaim_key: &XOnlyPublicKey,
-) -> CommitRevealResult<BitcoinAddress> {
-    commit(&deposit_data.serialize_to_vec(), revealer_key, reclaim_key)
 }
 
 /// Constructs a peg out payment address
@@ -108,23 +69,6 @@ pub fn withdrawal_request_commit(
         revealer_key,
         reclaim_key,
     )
-}
-
-/// Constructs a transaction that reveals the peg in payment address
-pub fn deposit_reveal_unsigned(
-    deposit_data: DepositData,
-    reveal_inputs: RevealInputs,
-    commit_amount: Amount,
-    peg_wallet_address: BitcoinAddress,
-) -> CommitRevealResult<Transaction> {
-    let mut tx = reveal(&deposit_data.serialize_to_vec(), reveal_inputs)?;
-
-    tx.output.push(TxOut {
-        value: (commit_amount - deposit_data.reveal_fee).to_sat(),
-        script_pubkey: peg_wallet_address.script_pubkey(),
-    });
-
-    Ok(tx)
 }
 
 /// Constructs a transaction that reveals the peg out payment address
