--- sbtc-core/src/operations/op_return/construction/mod.rs
@@ -1,12 +1,0 @@
-/*!
-Construction of OP_RETURN transactions for sBTC operations
-*/
-
-/// Module for deposit transaction construction
-pub mod deposit;
-/// Utilities for sBTC operation construction
-pub(crate) mod utils;
-/// Module for withdrawal fulfillment construction
-pub mod withdrawal_fulfillment;
-/// Module for withdrawal request construction
-pub mod withdrawal_request;
