--- sbtc-core/src/operations/op_return/construction/withdrawal_fulfillment.rs
@@ -1,1 +1,0 @@
-
