+++ sbtc-core/src/operations/op_return/mod.rs
@@ -1,2 +1,7 @@
-pub mod construction;
-pub mod parsing;
+/*!
+Primitives for sBTC OP_RETURN transactions
+*/
+pub mod deposit;
+pub mod utils;
+pub mod withdrawal_fulfillment;
+pub mod withdrawal_request;
