--- sbtc-core/src/operations/op_return/parsing/mod.rs
@@ -1,6 +1,0 @@
-/*!
-Parsing of OP_RETURN sBTC transactions
-*/
-pub mod deposit;
-pub mod withdrawal_fulfillment;
-pub mod withdrawal_request;
